Lift _run_injector_impl to BaseInjector as concrete template method
The Strategy pattern was incomplete while _run_injector_impl was abstract
and duplicated (~95% identical, ~35 lines) across both subclasses.

- Make _run_injector_impl a concrete template method in BaseInjector
- Remove duplicated implementations from ExhaustiveSEUInjector and
  StochasticSEUInjector (now inherited)
- Clean up unused imports in subclasses
- New injector types now only need _get_injection_indices to work
